commits and changes

+++ database/migrations/2021_03_05_092329_add_follower_id_to_users_table.php
@@ -14,7 +14,7 @@ class AddFollowerIdToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('follower_id')->nullable();
+            $table->bigInteger('follower_id')->nullable();
         });
     }
